+++ src/Internal/Protocol/Frame/BasicDeliver.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Thesis\Amqp\Internal\Protocol\Frame;
 
+use BcMath\Number;
 use Thesis\Amqp\Internal\Io;
 use Thesis\Amqp\Internal\Protocol\Frame;
 
@@ -12,12 +13,9 @@ use Thesis\Amqp\Internal\Protocol\Frame;
  */
 final readonly class BasicDeliver implements Frame
 {
-    /**
-     * @param non-negative-int $deliveryTag
-     */
     public function __construct(
         public string $consumerTag,
-        public int $deliveryTag,
+        public Number $deliveryTag,
         public bool $redelivered,
         public string $exchange,
         public string $routingKey,