+++ src/Internal/Protocol/Frame/ConnectionStartOk.php
@@ -7,7 +7,7 @@ namespace Thesis\Amqp\Internal\Protocol\Frame;
 use Thesis\Amqp\Internal\Io;
 use Thesis\Amqp\Internal\Protocol\Auth\Mechanism;
 use Thesis\Amqp\Internal\Protocol\Frame;
-use Thesis\Endian\endian;
+use Thesis\Endian\Order;
 
 /**
  * @internal
@@ -33,7 +33,7 @@ final readonly class ConnectionStartOk implements Frame
         $writer
             ->writeTable($this->clientProperties)
             ->writeString($this->auth->name())
-            ->reserve(endian::network->packUint32(...), $this->auth->write(...))
+            ->reserve(Order::Network->packUint32(...), $this->auth->write(...))
             ->writeString($this->locale);
     }
 }